Move table datasets to kart/tabular
