+++ kart/pk_generation.py
@@ -1,8 +1,8 @@
 import pygit2
 
 from .import_source import ImportSource
-from .dataset2 import Dataset2
-from .dataset2 import Dataset3
+from .tabular.dataset2 import Dataset2
+from .tabular.dataset2 import Dataset3
 from .tabular.schema import ColumnSchema
