+++ kart/repo_version.py
@@ -75,12 +75,12 @@ def dataset_class_for_version(version):
     """
     assert MIN_SUPPORTED_REPO_VERSION <= version <= MAX_SUPPORTED_REPO_VERSION
     if version == 2:
-        from kart.dataset2 import Dataset2
+        from kart.tabular.dataset2 import Dataset2
 
         return Dataset2
 
     if version == 3:
-        from kart.dataset3 import Dataset3
+        from kart.tabular.dataset3 import Dataset3
 
         return Dataset3
