+++ kart/upgrade/upgrade_v0.py
@@ -2,7 +2,7 @@ import functools
 import re
 
 from kart.geometry import normalise_gpkg_geom
-from kart.base_dataset import BaseDataset
+from kart.tabular.base_dataset import BaseDataset
 from kart.serialise_util import json_unpack
 from kart.sqlalchemy.adapter.gpkg import KartAdapter_GPKG
 from kart.utils import ungenerator
